refactor(pkg/trie): Split node handle from normal node
